modify domain field "qty" data type

+++ grails-app/domain/foodprint/Batch.groovy
@@ -23,7 +23,7 @@ class Batch  {
 	]
 
 	String name
-	Long expectQty = 0
+	double expectQty = 0
 	Date dueDate
 	Date manufactureDate
 	Date expirationDate
